requirements.txt
langchain-community
langchainhub
langgraph
mysql-connector-python
langchain_openai
langchain-groq
openpyxl
python-dotenv
pymysql
cryptography